optimize code structure
