+++ pkg/reconcile/apolloportal/apolloportal.go
@@ -1,13 +1,8 @@
 package apolloportal
 
-import (
-	"context"
-)
+type Apolloportal struct {
+}
 
-// Self updates this instance's self data. This should be the last item in the reconciliation, as it causes changes
-// making params.Instance obsolete. Default values should be set in the Defaulter webhook, this should only be used
-// for the Status, which can't be set by the defaulter.
-func Self(ctx context.Context, params Params) error {
-
-	return nil
+func NewApolloPortal() Apolloportal {
+	return Apolloportal{}
 }
