--- pkg/reconcile/apolloportal/ingress.go
@@ -1,11 +1,0 @@
-package apolloportal
-
-import (
-	"context"
-)
-
-// Ingresses reconciles the ingress(s) required for the instance in the current context.
-func Ingresses(ctx context.Context, params Params) error {
-
-	return nil
-}
