+++ pkg/utils/constant.go
@@ -1,0 +1,7 @@
+package utils
+
+// headless label is to differentiate the headless service from the clusterIP service.
+const (
+	HeadlessLabel  = "apollo.io/apollo-headless-service"
+	HeadlessExists = "Exists"
+)
